put dso/simplex noises in renamed directories
