+++ include/simplex/simplexnoise.h
