+++ other/dso_noises/simplexnoise1234.c
@@ -58,7 +58,7 @@ THE SOFTWARE.
  */
 
 // We don't need to include this. It does no harm, but no use either.
-#include "DSOnoises/simplexnoise1234.h"
+#include <dso_noises/simplexnoise1234.h>
 
 #define FASTFLOOR(x) (((x) > 0) ? ((int)x) : (((int)x) - 1))
