+++ other/dso_noises/srdnoise23.c
@@ -72,7 +72,7 @@ THE SOFTWARE.
 
 #include <math.h>
 
-#include "DSOnoises/srdnoise23.h" /* We strictly don't need this, but play nice. */
+#include "dso_noises/srdnoise23.h" /* We strictly don't need this, but play nice. */
 
 #define FASTFLOOR(x) (((x) > 0) ? ((int)x) : (((int)x) - 1))
